Encriptando password

+++ controllers/usercontroler.js
@@ -1,14 +1,36 @@
 import { User } from "../models/User.js";
+import bcrypt from 'bcryptjs'
 
 export class UserController {
   async create(req, res) {
-    const { body } = req;
-    const usuario = await User.create(body);
-    await usuario.save();
+    const { firstName, lastName, password, email } = req.body;
+
+    const usuario = new User({firstName,lastName,password,email})
+    
+   // email validation 
+
+   const emailExist =await User.findOne({where:{ email}})
+ 
+
+  
+    if(emailExist){
+      return  res.status(400).json({
+        msg:'este email ya fue registrado'
+      })
+      
+    }
+
+
+    // Encriptar password
+    const salt = bcrypt.genSaltSync()
+    usuario.password = bcrypt.hashSync(password,salt);
+
+    await usuario.save()
+
 
     res.status(200).json({
       msg: "Resgistrado corectamente",
-      usuario,
+     
     });
   }
 
@@ -36,20 +58,33 @@ export class UserController {
 
     await usuario.save();
     res.status(200).json({
-        msg:"Editado correctamente",
-        usuario
-    })
+      msg: "Editado correctamente",
+      usuario,
+    });
   }
 
   async Eliminar(req, res) {
-    const { id } = req.params;
-    await User.destroy({
-      where: {
+    try {
+      const { id } = req.params;
+      const usuario = await User.findByPk(id);
+
+      if (!usuario) {
+        res.status(404).json({
+          msg: "No se a encontrado el usuario",
+        });
+        return;
+      }
+      await User.destroy({
+        where: {
+          id,
+        },
+      });
+      res.status(200).json({
+        msg:"Se a eliminado Correctamente",
         id,
-      },
-    });
-    res.status(204).json({
-      msg: "Se a eliminado Correctamente",
-    });
+      });
+    } catch (error) {
+      console.log(error);
+    }
   }
 }
